# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FANAH Fragrances Decants Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add a decant variant to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variantId, quantity]
              properties:
                variantId:
                  type: integer
                quantity:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update quantity of a cart line
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove a line from the cart
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
components:
  schemas:
...
    CartItem:
      type: object
      required: [id, variantId, productId, productName, productSlug, brand, sizeMl, imageUrl, unitPriceCents, quantity, lineTotalCents]
      properties:
        id:
          type: integer
        variantId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productSlug:
          type: string
        brand:
          type: string
        sizeMl:
          type: integer
        imageUrl:
          type: string
        unitPriceCents:
          type: integer
        quantity:
          type: integer
        lineTotalCents:
          type: integer
    Cart:
      type: object
      required: [items, subtotalCents, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotalCents:
          type: integer
        itemCount:
          type: integer
...